Fixed Rendering Issue
RenderTickEvent registered wrong.
