+++ src/main/java/zdoctor/hp/client/renderer/HelperRender.java
@@ -1,0 +1,8 @@
+package zdoctor.hp.client.renderer;
+
+public class HelperRender {
+    public static double partialTicks = 0;
+    public static double calcRenderPos(double pos, double prevPos) {
+        return (double) (prevPos+(pos-prevPos)*HelperRender.partialTicks);
+    }
+}
